fix in nutritionix api get details

+++ PregnancyHelper/app/src/main/java/hr/ferit/helenaborzan/pregnancyhelper/model/AltMeasure.kt
@@ -2,7 +2,7 @@ package hr.ferit.helenaborzan.pregnancyhelper.model
 
 data class AltMeasure(
     val measure: String?,
-    val qty: Int?,
+    val qty: Double?,
     val seq: Int?,
     val serving_weight: Double?
 )
